+++ shibuya-kafka-adapter/test/Main.hs
@@ -3,7 +3,6 @@ module Main (main) where
 import Shibuya.Adapter.Kafka.AdapterTest qualified as AdapterTest
 import Shibuya.Adapter.Kafka.ConvertTest qualified as ConvertTest
 import Shibuya.Adapter.Kafka.IntegrationTest qualified as IntegrationTest
-import Shibuya.Adapter.Kafka.TracingTest qualified as TracingTest
 import Test.Tasty (defaultMain, testGroup)
 
 main :: IO ()
@@ -14,5 +13,4 @@ main =
             [ AdapterTest.tests
             , ConvertTest.tests
             , IntegrationTest.tests
-            , TracingTest.tests
             ]